Move benchmarks tests
