+++ ksmt-test/src/test/kotlin/org/ksmt/test/benchmarks/Z3ConverterBenchmark.kt
@@ -1,4 +1,4 @@
-package org.ksmt.test
+package org.ksmt.test.benchmarks
 
 import com.microsoft.z3.Context
 import kotlinx.coroutines.async
